+++ CHANGELOG.md
@@ -6,11 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
 # Added
 
 - ✨Add image attachments with access control
-
-## [Unreleased]
+- ✨(frontend) Upload image to a document
 
 
 ## [1.2.1] - 2024-08-23